move shared types out
