+++ packages/interface/README.md
@@ -1,0 +1,3 @@
+# `@window-splitter/interface`
+
+Shared component interface for window splitter components.
