+++ packages/interface/src/index.ts
@@ -1,0 +1,101 @@
+import {
+  Constraints,
+  GroupMachineContextValue,
+  OnResizeCallback,
+  PanelData,
+  PixelUnit,
+  Unit,
+} from "@window-splitter/state";
+
+export interface PanelGroupHandle {
+  /** The id of the group */
+  getId: () => string;
+  /** Get the sizes of all the items in the layout as pixels */
+  getPixelSizes: () => Array<number>;
+  /** Get the sizes of all the items in the layout as percentages of the group size */
+  getPercentageSizes: () => Array<number>;
+  /**
+   * Set the size of all the items in the layout.
+   * This just calls `setSize` on each item. It is up to
+   * you to make sure the sizes make sense.
+   *
+   * NOTE: Setting handle sizes will do nothing.
+   */
+  setSizes: (items: Array<Unit>) => void;
+  /** Get the template for the group in pixels. Useful for testing */
+  getTemplate: () => string;
+  getState: () => "idle" | "dragging";
+}
+
+export interface SharedPanelGroupProps
+  extends Pick<GroupMachineContextValue, "orientation" | "autosaveStrategy"> {
+  /** Persisted state to initialized the machine with */
+  snapshot?: GroupMachineContextValue;
+  /** An id to use for autosaving the layout */
+  autosaveId?: string;
+}
+
+export interface PanelHandle {
+  /** Collapse the panel */
+  collapse: () => void;
+  /** Returns true if the panel is collapsed */
+  isCollapsed: () => boolean;
+  /** Expand the panel */
+  expand: () => void;
+  /** Returns true if the panel is expanded */
+  isExpanded: () => boolean;
+  /** The id of the panel */
+  getId: () => string;
+  /** Get the size of the panel in pixels */
+  getPixelSize: () => number;
+  /** Get percentage of the panel relative to the group */
+  getPercentageSize: () => number;
+  /**
+   * Set the size of the panel in pixels.
+   *
+   * This will be clamped to the min/max values of the panel.
+   * If you want the panel to collapse/expand you should use the
+   * expand/collapse methods.
+   */
+  setSize: (size: Unit) => void;
+}
+export interface SharedPanelProps<CollapsValue>
+  extends Constraints<Unit>,
+    Pick<PanelData, "collapseAnimation"> {
+  /** Callback called when the panel is resized */
+  onResize?: OnResizeCallback;
+  /**
+   * __CONTROLLED COMPONENT__
+   *
+   * If this prop is used it will be used as the source of truth for the collapsed state.
+   * It should be used in conjunction with the `onCollapseChange` prop.
+   *
+   * Use this if you want full control over the collapsed state. When trying to
+   * collapse a panel it will defer to onCollapseChange to determine if it should
+   * be collapsed.
+   */
+  collapsed?: CollapsValue;
+  /**
+   * __CONTROLLED COMPONENT__
+   *
+   * A callback called with the new desired collapsed state. If paired w
+   * with the `collapsed` prop this will be used to control the collapsed state.
+   *
+   * Otherwise this will just be called with the new collapsed state so you can
+   * use it to update your own state.
+   */
+  onCollapseChange?: (isCollapsed: boolean) => void;
+}
+
+export interface SharedPanelResizerProps extends Pick<PanelHandle, "setSize"> {
+  /** If the handle is disabled */
+  disabled?: boolean;
+  /** The size of the handle */
+  size?: PixelUnit;
+  /** Called when the user starts dragging the handle */
+  onDragStart?: () => void;
+  /** Called when the user drags the handle */
+  onDrag?: () => void;
+  /** Called when the user stops dragging the handle */
+  onDragEnd?: () => void;
+}
