+++ packages/solid/package.json
@@ -56,6 +56,7 @@
     "solid-js": ">=1.x"
   },
   "dependencies": {
+    "@window-splitter/interface": "workspace:*",
     "@window-splitter/state": "workspace:*"
   },
   "tshy": {
